# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "narrator",
    "description": "Self-hosted TTS book reader. One global session: the server holds one book, one chapter and one playhead across every client, which is why anything cacheable is scoped by ?book= and positions can be written by name. No auth, no TLS, wildcard CORS - it assumes a private machine reached over Tailscale.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/books/delete": {
      "post": {
        "tags": [
          "library"
        ],
        "operationId": "delete_book",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok2"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/books/upload": {
      "post": {
        "tags": [
          "library"
        ],
        "operationId": "upload",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/epub+zip": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookFile"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/library": {
      "get": {
        "tags": [
          "library"
        ],
        "summary": "Every book on this box, with what is rendered and packed for each.",
        "operationId": "library",
        "parameters": [
          {
            "name": "book",
            "in": "query",
            "description": "One book's cache key. Unknown keys answer with an empty list rather than\na 404: this endpoint describes what is here, and \"not here\" is an answer.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "chapters",
            "in": "query",
            "description": "Include the per-chapter rows. Only meaningful alongside `book`; it is\nignored for the all-books answer, which would otherwise be enormous.",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "description": "Every error body in this API: `{\"error\": \"...\"}`, and nothing else. The\nreader's `get()` reads exactly this field.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "BookFile": {
        "type": "object",
        "required": [
          "path",
          "name",
          "mb"
        ],
        "properties": {
          "mb": {
            "type": "number",
            "format": "double"
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string",
            "description": "Absolute path on the server — what `/api/load` takes back."
          }
        }
      },
...
      "ChapterIndexRow": {
        "type": "object",
        "description": "A cached filesystem scan of one chapter. **Not truth** — see the module doc.\n\nThe field names are [`crate::api::chapters::ChapterRow`]'s on purpose: this\ntable exists to answer that endpoint for a book the session has not loaded,\nand a row that has to be renamed on the way out is a row that will eventually\nbe renamed wrongly.",
        "required": [
          "chapter",
          "n",
          "rendered",
          "m4a",
          "title",
          "scanned_ms"
        ],
        "properties": {
          "bytes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "chapter": {
            "type": "integer",
            "minimum": 0
          },
          "duration": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "est_s": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Estimated spoken seconds, from the chunker's character counts."
          },
          "m4a": {
            "type": "boolean"
          },
          "n": {
            "type": "integer",
            "description": "Chunks in the plan.",
            "minimum": 0
          },
          "rendered": {
            "type": "integer",
            "description": "Chunk wavs that were on disk **when the scan ran**.",
            "minimum": 0
          },
          "scanned_ms": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "DeleteBody": {
        "type": "object",
        "required": [
          "path",
          "confirm"
        ],
        "properties": {
          "confirm": {
            "type": "string",
            "description": "Must exactly match the EPUB filename to confirm permanent deletion."
          },
          "path": {
            "type": "string"
          }
        }
      },
...
      "LibraryBook": {
        "type": "object",
        "description": "One book on this box.\n\nThe counts are aggregates of that book's cached scan rows; `chapters` is what\nthe `book` table records, so a book whose plan is on disk but which has never\nbeen scanned still says how long it is.",
        "required": [
          "key",
          "name",
          "path",
          "title",
          "chapters",
          "last_open_ms",
          "rendered_chapters",
          "packed_chapters",
          "packed_bytes",
          "total_chunks",
          "rendered_chunks",
          "est_min",
          "position",
          "loaded"
        ],
        "properties": {
          "chapter_index": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/ChapterIndexRow"
            },
            "description": "The per-chapter rows, present only for `?book=<key>&chapters=true`.\n\nDeliberately absent from the all-books answer: *Lord of Mysteries* alone\nis 1433 of these, and a library response that carried them for every book\nwould be megabytes over a tunnel to answer a question about shelves."
          },
          "chapters": {
            "type": "integer",
            "minimum": 0
          },
          "est_min": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "The whole book's spoken length, from the chunker's character counts.\nNull only when nothing has been scanned yet."
          },
          "key": {
            "type": "string",
            "description": "The cache directory name — what `?book=` takes everywhere else."
          },
          "last_open_ms": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "When this book was last opened, or null for one that has only ever been\nfound on disk."
          },
          "loaded": {
            "type": "boolean",
            "description": "Is this the book the session is holding? The one field here that is a\nlive fact rather than a scan result."
          },
          "name": {
            "type": "string",
            "description": "The epub file name, which is what positions are keyed by."
          },
          "packed_bytes": {
            "type": "integer",
            "format": "int64",
            "description": "What those weigh, measured — not estimated.",
            "minimum": 0
          },
          "packed_chapters": {
            "type": "integer",
            "description": "Chapters with a *usable* packed m4a. A manifest that disagrees with the\nplan is not one; see [`crate::library::scan_book`].",
            "minimum": 0
          },
          "path": {
            "type": "string",
            "description": "Absolute path on the server, or empty for a book whose audio is here and\nwhose source file is not."
          },
          "position": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StampedPosition",
                "description": "Where this book was left, by **file name** — the store's newest record\nacross every device, falling back to the vault's own map for a book that\npredates the store. Null for one nobody has opened."
              }
            ]
          },
          "rendered_chapters": {
            "type": "integer",
            "description": "Chapters every chunk of which was on disk at the scan. A chapter with no\nchunks at all does not count: \"nothing to render\" is not \"rendered\".",
            "minimum": 0
          },
          "rendered_chunks": {
            "type": "integer",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "total_chunks": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "LibraryResult": {
        "type": "object",
        "required": [
          "books",
          "scanned_ms"
        ],
        "properties": {
          "books": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LibraryBook"
            }
          },
          "scanned_ms": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "The **oldest** scan in this answer, so the reader can say how fresh it\nis — and so one book that has not been walked since yesterday cannot hide\nbehind eleven that were walked a minute ago. Null when nothing in the\nanswer has ever been scanned."
          }
        }
      },
...
      "Ok2": {
        "type": "object",
        "description": "`{\"ok\": true}` plus whatever else a handler wants to say.",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      },
...
      "StampedPosition": {
        "type": "object",
        "description": "A position as the *API* hands it back: the record above plus an unambiguous\ninstant.\n\n`updated` is a naive local stamp with no zone, because that is what the\npython server writes into the vault and the file has to keep matching. A\nbrowser cannot resolve it — a container without `/etc/localtime` stamps UTC\nwhile the same reader's other positions are local, and the reader is left\nguessing which of two positions is newer. So the API carries the instant as\nwell, resolved here where the server's own zone is known.",
        "required": [
          "chapter",
          "chunk",
          "chapter_title",
          "chunks_total",
          "chapters_total",
          "updated"
        ],
        "properties": {
          "chapter": {
            "type": "integer",
            "format": "int64"
          },
          "chapter_title": {
            "type": "string"
          },
          "chapters_total": {
            "type": "integer",
            "format": "int64"
          },
          "chunk": {
            "type": "integer",
            "format": "int64"
          },
          "chunks_total": {
            "type": "integer",
            "format": "int64"
          },
          "updated": {
            "type": "string",
            "description": "Naive local ISO, byte-identical to the vault record."
          },
          "updated_ms": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "The same moment in epoch milliseconds. Null only if `updated` cannot be\nparsed at all, which means it was not written by a narrator."
          }
        }
      },
...
    }
  },
...
}